command line requirement changes
If there was a CustomerID given through commandline, it was possible to run the powershell with an empty RegistrationToken. It's unlogic that you provide a CustomerID through commandline and the registration token through PartnerConfig.xml.
If one is given and not the other it now quits with an error instead of trying to run the powershell script